Added a block switch, updated the fade block button and page control as well as paging scroll view. The new selected content button with animations and added the MAXViewController for automatic handling of the keyboard.

+++ ReusableComponentsApp/MAXBlockView.m
@@ -133,7 +133,7 @@
     
     NSMutableArray *theViews = [NSMutableArray array];
     
-    int numColumns = [self.datasource numColumnsInBlockView:self inRow:theRow];
+    NSInteger numColumns = [self.datasource numColumnsInBlockView:self inRow:theRow];
     
     float separatorWidth = [self.delegate widthForVerticalSeparatorForRow:theRow];
